🔧 Address all code review feedback and improve code quality
## Security & Performance Fixes
- **Fix mass assignment vulnerability**: Replace body spreading with explicit field validation in character update endpoint
- **Improve performance**: Replace delete operators with undefined assignments and use proper typing

## Code Quality Improvements
- **Extract reusable component**: Create CharacterSidebarSection to eliminate 70+ lines of duplicated code
- **Add middleware**: Implement verifyProjectAccess middleware to reduce repeated validation logic across 5 endpoints
- **Fix form state management**: Add useEffect to properly reset character dialog when props change
- **Remove redundant code**: Clean up unnecessary API method checks and unused imports

## Type Safety & Linting
- **Improve typing**: Replace `Record<string, any>` with specific union type for better type safety
- **Fix compilation errors**: Resolve TypeScript issues with optional method calls using Required<> utility
- **Clean up imports**: Remove unused lucide-react icons and fix import organization
- **Address dependency warnings**: Fix useEffect dependencies to prevent unnecessary re-renders

## Technical Debt Reduction
- **Reduce code duplication**: From 140+ lines of duplicated sidebar logic to single reusable component
- **Centralize validation**: Move project access checks to middleware, removing 80+ lines of repeated code
- **Improve maintainability**: Consistent patterns across all character CRUD endpoints

All quality checks now pass: TypeScript compilation ✅, Linting ✅, Build process ✅
