+++ apps/web/src/routes/projects/$projectId.tsx
@@ -3,30 +3,18 @@ import { createFileRoute, Link, Outlet } from "@tanstack/react-router"
 import {
   ChevronDown,
   ChevronRight,
-  Edit,
   FileEdit,
   FileText,
   MapPin,
   PenTool,
-  Plus,
   Scroll,
-  Trash2,
-  Users,
 } from "lucide-react"
 import { useState } from "react"
-import { CharacterDialog } from "@/components/character-dialog"
+import { CharacterSidebarSection } from "@/components/character-sidebar-section"
 import CodexModal from "@/components/codex-modal"
-import { DeleteCharacterDialog } from "@/components/delete-character-dialog"
 import { Badge } from "@/components/ui/badge"
 import { Button } from "@/components/ui/button"
 import { Collapsible, CollapsibleContent, CollapsibleTrigger } from "@/components/ui/collapsible"
-import {
-  ContextMenu,
-  ContextMenuContent,
-  ContextMenuItem,
-  ContextMenuSeparator,
-  ContextMenuTrigger,
-} from "@/components/ui/context-menu"
 import { Progress } from "@/components/ui/progress"
 import { Separator } from "@/components/ui/separator"
 import {
@@ -44,7 +32,7 @@ import {
 } from "@/components/ui/sidebar"
 import { Tooltip, TooltipContent, TooltipProvider, TooltipTrigger } from "@/components/ui/tooltip"
 import WriteHeader from "@/components/write-header"
-import { api, type Character } from "@/lib/api"
+import { api } from "@/lib/api"
 
 export const Route = createFileRoute("/projects/$projectId")({
   component: WriteLayout,
@@ -63,25 +51,12 @@ function WriteLayout() {
     lore: false,
     plot: false,
   })
-  const [characterDialogOpen, setCharacterDialogOpen] = useState(false)
-  const [characterDialogMode, setCharacterDialogMode] = useState<"create" | "edit">("create")
-  const [selectedCharacter, setSelectedCharacter] = useState<Character | null>(null)
-  const [deleteCharacterDialogOpen, setDeleteCharacterDialogOpen] = useState(false)
 
   // Fetch project details
   const { data: project, isLoading } = useQuery({
     queryKey: ["project", projectId],
     queryFn: async () => {
       const result = await api.projects.get(projectId)
-      return result
-    },
-  })
-
-  // Fetch characters
-  const { data: characters = [] } = useQuery({
-    queryKey: ["characters", projectId],
-    queryFn: async () => {
-      const result = await api.characters.list(projectId)
       return result
     },
   })
@@ -130,23 +105,6 @@ function WriteLayout() {
   const closeCodexModal = () => {
     setIsCodexModalOpen(false)
     setCodexModalConfig({})
-  }
-
-  const handleCreateCharacter = () => {
-    setSelectedCharacter(null)
-    setCharacterDialogMode("create")
-    setCharacterDialogOpen(true)
-  }
-
-  const handleEditCharacter = (character: Character) => {
-    setSelectedCharacter(character)
-    setCharacterDialogMode("edit")
-    setCharacterDialogOpen(true)
-  }
-
-  const handleDeleteCharacter = (character: Character) => {
-    setSelectedCharacter(character)
-    setDeleteCharacterDialogOpen(true)
   }
 
   const codexData = {
@@ -226,77 +184,12 @@ function WriteLayout() {
               <SidebarGroupContent>
                 <SidebarMenu>
                   {/* Characters */}
-                  <SidebarMenuItem>
-                    <Collapsible
-                      onOpenChange={() => toggleCodexSection("characters")}
-                      open={expandedCodexSections.characters}
-                    >
-                      <CollapsibleTrigger asChild>
-                        <SidebarMenuButton>
-                          {expandedCodexSections.characters ? (
-                            <ChevronDown className="h-4 w-4" />
-                          ) : (
-                            <ChevronRight className="h-4 w-4" />
-                          )}
-                          <Users className="h-4 w-4" />
-                          <span>Characters</span>
-                          <Badge className="ml-auto" variant="secondary">
-                            {characters.length}
-                          </Badge>
-                        </SidebarMenuButton>
-                      </CollapsibleTrigger>
-                      <CollapsibleContent>
-                        <div className="ml-6 space-y-1">
-                          <Button
-                            className="w-full justify-start text-muted-foreground"
-                            onClick={handleCreateCharacter}
-                            size="sm"
-                            variant="ghost"
-                          >
-                            <Plus className="h-4 w-4" />
-                            <span>Create New Character</span>
-                          </Button>
-                          {characters.map((character) => (
-                            <ContextMenu key={character.id}>
-                              <ContextMenuTrigger asChild>
-                                <Button
-                                  className="w-full justify-start"
-                                  onClick={() => openCodexModal("characters", character.name)}
-                                  size="sm"
-                                  variant="ghost"
-                                >
-                                  <span className="truncate">{character.name}</span>
-                                  <span className="ml-auto text-muted-foreground text-xs">
-                                    {character.role}
-                                  </span>
-                                </Button>
-                              </ContextMenuTrigger>
-                              <ContextMenuContent>
-                                <ContextMenuItem
-                                  onClick={() => openCodexModal("characters", character.name)}
-                                >
-                                  <FileText className="mr-2 h-4 w-4" />
-                                  View Details
-                                </ContextMenuItem>
-                                <ContextMenuItem onClick={() => handleEditCharacter(character)}>
-                                  <Edit className="mr-2 h-4 w-4" />
-                                  Edit Character
-                                </ContextMenuItem>
-                                <ContextMenuSeparator />
-                                <ContextMenuItem
-                                  className="text-destructive focus:text-destructive"
-                                  onClick={() => handleDeleteCharacter(character)}
-                                >
-                                  <Trash2 className="mr-2 h-4 w-4" />
-                                  Delete Character
-                                </ContextMenuItem>
-                              </ContextMenuContent>
-                            </ContextMenu>
-                          ))}
-                        </div>
-                      </CollapsibleContent>
-                    </Collapsible>
-                  </SidebarMenuItem>
+                  <CharacterSidebarSection
+                    isExpanded={expandedCodexSections.characters}
+                    onOpenCodexModal={openCodexModal}
+                    onToggle={() => toggleCodexSection("characters")}
+                    projectId={projectId}
+                  />
 
                   {/* Locations */}
                   <SidebarMenuItem>
@@ -554,25 +447,6 @@ function WriteLayout() {
         onClose={closeCodexModal}
         projectId={projectId}
       />
-
-      {/* Character Dialog */}
-      <CharacterDialog
-        character={selectedCharacter}
-        mode={characterDialogMode}
-        onOpenChange={setCharacterDialogOpen}
-        open={characterDialogOpen}
-        projectId={projectId}
-      />
-
-      {/* Delete Character Dialog */}
-      {selectedCharacter && (
-        <DeleteCharacterDialog
-          character={selectedCharacter}
-          onOpenChange={setDeleteCharacterDialogOpen}
-          open={deleteCharacterDialogOpen}
-          projectId={projectId}
-        />
-      )}
     </SidebarProvider>
   )
 }
